Add the Octopus by Shinji Minamiyama
This program is from [micomBASIC 1985-08, pp. 154-155]. It contains
one BASIC program, `octopus.bas`.  This is included on the
self-booting disk image `octopus.dsk`.

The code was transcribed by @eientei from the applesaucefdc Discord
server.

The `README` explains how the game works.  As the machine language
routines are a subset of those for _The Active Balls_, there is no
disassembly.

[micomBASIC 1985-08, pp. 154-155]: https://archive.org/details/micomBASIC-1985-08/page/154/mode/1up